+++ R/zzz.R
@@ -1,2 +1,2 @@
 globalVariables(c("n", "y", "x", "predicted", "lwr_95", "upr_95", "frequency", ".SD", ".N", ".", "..density..",
-                  "avg_claimsize", "avg_premium", ":=", "difference", "maanden", "maanden_tot", "id"))
+                  "avg_claimsize", "avg_premium", ":=", "difference", "maanden", "maanden_tot", "id", "..count.."))